scripts: build flashable zip directly

+++ scripts/internal/create_target_files_zip.sh
@@ -5,8 +5,6 @@
 # [
 source "$SRC_DIR/scripts/utils/install_utils.sh" || exit 1
 
-trap 'rm -rf "$TMP_DIR"' EXIT INT
-
 # https://android.googlesource.com/platform/build/+/refs/tags/android-15.0.0_r1/tools/releasetools/build_super_image.py#72
 BUILD_SUPER_EMPTY()
 {
@@ -117,12 +115,18 @@ GET_SUPER_GROUP_SIZE()
 }
 # ]
 
-if [ "$#" != "1" ]; then
-    echo "Usage: create_target_files_zip <output>" >&2
+if [[ "$#" == "2" ]] && [[ "$1" == "--directory" ]]; then
+    OUTPUT_MODE="directory"
+    OUTPUT_FILE="$2"
+    TMP_DIR="$OUTPUT_FILE"
+elif [[ "$#" == "1" ]]; then
+    OUTPUT_MODE="zip"
+    OUTPUT_FILE="$1"
+    trap 'rm -rf "$TMP_DIR"' EXIT INT
+else
+    echo "Usage: create_target_files_zip [--directory] <output>" >&2
     exit 1
 fi
-
-OUTPUT_FILE="$1"
 
 [ -d "$TMP_DIR" ] && rm -rf "$TMP_DIR"
 mkdir -p "$TMP_DIR"
@@ -168,6 +172,10 @@ fi
 
 LOG "- Generating build_info.txt"
 GENERATE_BUILD_INFO
+
+if [[ "$OUTPUT_MODE" == "directory" ]]; then
+    exit 0
+fi
 
 LOG "- Creating zip"
 rm -f "$OUTPUT_FILE"
